+++ source/extensions/isaacsim.core.prims/python/impl/cloth_prim.py
@@ -16,8 +16,10 @@
 from typing import Optional, Union
 
 import carb
+import carb.eventdispatcher
 import numpy as np
 import omni.kit.app
+import omni.timeline
 from isaacsim.core.deprecation_manager import import_module
 from pxr import PhysxSchema, UsdPhysics, Vt
 
@@ -149,8 +151,10 @@ class ClothPrim(XFormPrim):
                 self.set_spring_dampings(spring_dampings)
 
         timeline = omni.timeline.get_timeline_interface()
-        self._invalidate_physics_handle_event = timeline.get_timeline_event_stream().create_subscription_to_pop(
-            self._invalidate_physics_handle_callback
+        self._invalidate_physics_handle_event = carb.eventdispatcher.get_eventdispatcher().observe_event(
+            event_name=omni.timeline.GLOBAL_EVENT_STOP,
+            on_event=self._invalidate_physics_handle_callback,
+            observer_name="isaacsim.core.prims.ClothPrim.initialize._invalidate_physics_handle_callback",
         )
 
     """
@@ -212,8 +216,7 @@ class ClothPrim(XFormPrim):
         return
 
     def _invalidate_physics_handle_callback(self, event):
-        if event.type == int(omni.timeline.TimelineEventType.STOP):
-            self._physics_view = None
+        self._physics_view = None
         return
 
     def _apply_cloth_auto_api(self, index):