+++ source/extensions/isaacsim.sensors.physx.examples/isaacsim/sensors/physx/examples/lightbeam_sensor.py
@@ -16,11 +16,13 @@ import asyncio
 import weakref
 
 import carb
+import carb.eventdispatcher
 import omni
 import omni.graph.core as og
 import omni.kit.commands
 import omni.physics.core
 import omni.ui as ui
+import omni.usd
 from isaacsim.core.utils.prims import delete_prim, get_prim_at_path
 from isaacsim.core.utils.viewports import set_camera_view
 from isaacsim.examples.browser import get_instance as get_browser_instance
@@ -51,9 +53,9 @@ class LightBeamSensorDemo(omni.ext.IExt):
     def build_window(self):
         pass
 
-    def _on_stage_event(self, event):
-        if event.type == int(omni.usd.StageEventType.CLOSED):
-            self.on_closed()
+    def _on_stage_closed(self, event):
+        """Stage closed event callback."""
+        self.on_closed()
 
     def build_ui(self):
         with ui.VStack(spacing=5, height=0):
@@ -245,9 +247,11 @@ class LightBeamSensorDemo(omni.ext.IExt):
         # we want to make sure we can see the sensor we made, so we set the camera position and look target
         set_camera_view(eye=[-5.00, 5.00, 3.50], target=[0.0, 0.0, 0.0], camera_prim_path="/OmniverseKit_Persp")
 
-        self._events = omni.usd.get_context().get_stage_event_stream()
-        self._stage_event_subscription = self._events.create_subscription_to_pop(
-            self._on_stage_event, name="LightBeam Sensor Sample Stage Watch"
+        self._usd_context = omni.usd.get_context()
+        self._stage_event_subscription = carb.eventdispatcher.get_eventdispatcher().observe_event(
+            event_name=self._usd_context.stage_event_name(omni.usd.StageEventType.CLOSED),
+            on_event=self._on_stage_closed,
+            observer_name="isaacsim.sensors.physx.examples.lightbeam_sensor._on_stage_closed",
         )
 
         (action_graph, new_nodes, _, _) = og.Controller.edit(