+++ source/extensions/isaacsim.gui.sensors.icon/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [2.0.6] - 2025-12-05
+### Changed
+- Migrate to Events 2.0.
 
 ## [2.0.5] - 2025-12-01
 ### Changed