+++ source/extensions/isaacsim.robot_setup.grasp_editor/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [2.1.2] - 2025-12-05
+### Changed
+- Migrate to Events 2.0.
 
 ## [2.1.1] - 2025-10-27
 ### Changed