+++ source/extensions/isaacsim.sensors.rtx/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [15.11.3] - 2025-12-05
+### Changed
+- Migrate to Events 2.0.
+
 ## [15.11.2] - 2025-12-03
 ### Changed
 - Remove TODOs.